+++ script/sql/psql/upgrade_database_from_v1_to_v2.sql
@@ -1,4 +1,4 @@
 -- Update ArchiveFile
 ALTER TABLE ArchiveFile ALTER name TYPE TEXT;
-ALTER TABLE ArchiveFile ADD blockNumber INTEGER CHECK (blockNumber >= 0);
+ALTER TABLE ArchiveFile ADD blockNumber BIGINT CHECK (blockNumber >= 0);
